feat: add polyfill so as not to pollute global
